+++ libcrux-ml-kem/Cargo.toml
@@ -13,8 +13,7 @@ exclude = ["/tests", "/implementation_notes.pdf"]
 rand_core = { version = "0.6" }
 libcrux-platform = { version = "0.0.2-pre.2", path = "../sys/platform" }
 libcrux-sha3 = { version = "0.0.2-pre.2", path = "../libcrux-sha3" }
-# libcrux-polynomials-aarch64 = { path = "../polynomials-aarch64/" }
-libcrux-polynomials-avx2 = { path = "../polynomials-avx2/" }
+libcrux-polynomials = { path = "../polynomials/" }
 
 # This is only required for verification.
 # The hax config is set by the hax toolchain.
@@ -23,8 +22,8 @@ hax-lib = { git = "https://github.com/hacspec/hax/" }
 
 [features]
 default = ["simd256"]
-simd128 = ["libcrux-sha3/simd128"]
-simd256 = ["libcrux-sha3/simd256"]
+simd128 = ["libcrux-sha3/simd128", "libcrux-polynomials/simd128"]
+simd256 = ["libcrux-sha3/simd256", "libcrux-polynomials/simd256"]
 
 [dev-dependencies]
 rand = { version = "0.8" }
