+++ libcrux-ml-kem/src/lib.rs
@@ -12,8 +12,6 @@ pub(crate) mod constants;
 /// Helpers for verification and extraction
 mod helper;
 
-mod arithmetic;
-mod compress;
 mod constant_time_ops;
 mod hash_functions;
 mod ind_cpa;
@@ -22,7 +20,6 @@ mod ntt;
 mod polynomial;
 mod sampling;
 mod serialize;
-mod simd;
 mod types;
 
 // Variants
@@ -30,6 +27,7 @@ pub mod kyber1024;
 pub mod kyber512;
 pub mod kyber768;
 
+use libcrux_polynomials::{PortableVector, SIMD256Vector};
 pub use types::{MlKemCiphertext, MlKemKeyPair, MlKemPrivateKey, MlKemPublicKey};
 
 // TODO: We should make this an actual type as opposed to alias so we can enforce
@@ -80,17 +78,37 @@ pub(crate) fn validate_public_key<
 >(
     public_key: &[u8; PUBLIC_KEY_SIZE],
 ) -> bool {
-    let deserialized_pk = deserialize_ring_elements_reduced::<PUBLIC_KEY_SIZE, K>(
-        &public_key[..RANKED_BYTES_PER_RING_ELEMENT],
-    );
-
-    let public_key_serialized =
-        serialize_public_key::<K, RANKED_BYTES_PER_RING_ELEMENT, PUBLIC_KEY_SIZE>(
+    if libcrux_platform::simd256_support() {
+        let deserialized_pk = deserialize_ring_elements_reduced::<PUBLIC_KEY_SIZE, K, SIMD256Vector>(
+            &public_key[..RANKED_BYTES_PER_RING_ELEMENT],
+        );
+        let public_key_serialized = serialize_public_key::<
+            K,
+            RANKED_BYTES_PER_RING_ELEMENT,
+            PUBLIC_KEY_SIZE,
+            SIMD256Vector,
+        >(
             deserialized_pk,
             &public_key[RANKED_BYTES_PER_RING_ELEMENT..],
         );
 
-    *public_key == public_key_serialized
+        *public_key == public_key_serialized
+    } else {
+        let deserialized_pk = deserialize_ring_elements_reduced::<PUBLIC_KEY_SIZE, K, PortableVector>(
+            &public_key[..RANKED_BYTES_PER_RING_ELEMENT],
+        );
+        let public_key_serialized = serialize_public_key::<
+            K,
+            RANKED_BYTES_PER_RING_ELEMENT,
+            PUBLIC_KEY_SIZE,
+            PortableVector,
+        >(
+            deserialized_pk,
+            &public_key[RANKED_BYTES_PER_RING_ELEMENT..],
+        );
+
+        *public_key == public_key_serialized
+    }
 }
 
 pub(crate) fn generate_keypair<
@@ -107,21 +125,47 @@ pub(crate) fn generate_keypair<
     let ind_cpa_keypair_randomness = &randomness[0..CPA_PKE_KEY_GENERATION_SEED_SIZE];
     let implicit_rejection_value = &randomness[CPA_PKE_KEY_GENERATION_SEED_SIZE..];
 
-    let (ind_cpa_private_key, public_key) = ind_cpa::generate_keypair::<
-        K,
-        CPA_PRIVATE_KEY_SIZE,
-        PUBLIC_KEY_SIZE,
-        BYTES_PER_RING_ELEMENT,
-        ETA1,
-        ETA1_RANDOMNESS_SIZE,
-    >(ind_cpa_keypair_randomness);
+    // Runtime feature detection.
+    // TODO:
+    // - add aarch64
+    // - fallbacks
+    // - compilation on all platforms
+    // - pull out inner function
+    if libcrux_platform::simd256_support() {
+        let (ind_cpa_private_key, public_key) = ind_cpa::generate_keypair::<
+            K,
+            CPA_PRIVATE_KEY_SIZE,
+            PUBLIC_KEY_SIZE,
+            BYTES_PER_RING_ELEMENT,
+            ETA1,
+            ETA1_RANDOMNESS_SIZE,
+            SIMD256Vector,
+        >(ind_cpa_keypair_randomness);
 
-    let secret_key_serialized =
-        serialize_kem_secret_key(&ind_cpa_private_key, &public_key, implicit_rejection_value);
-    let private_key: MlKemPrivateKey<PRIVATE_KEY_SIZE> =
-        MlKemPrivateKey::from(secret_key_serialized);
+        let secret_key_serialized =
+            serialize_kem_secret_key(&ind_cpa_private_key, &public_key, implicit_rejection_value);
+        let private_key: MlKemPrivateKey<PRIVATE_KEY_SIZE> =
+            MlKemPrivateKey::from(secret_key_serialized);
 
-    MlKemKeyPair::from(private_key, public_key.into())
+        MlKemKeyPair::from(private_key, public_key.into())
+    } else {
+        let (ind_cpa_private_key, public_key) = ind_cpa::generate_keypair::<
+            K,
+            CPA_PRIVATE_KEY_SIZE,
+            PUBLIC_KEY_SIZE,
+            BYTES_PER_RING_ELEMENT,
+            ETA1,
+            ETA1_RANDOMNESS_SIZE,
+            PortableVector,
+        >(ind_cpa_keypair_randomness);
+
+        let secret_key_serialized =
+            serialize_kem_secret_key(&ind_cpa_private_key, &public_key, implicit_rejection_value);
+        let private_key: MlKemPrivateKey<PRIVATE_KEY_SIZE> =
+            MlKemPrivateKey::from(secret_key_serialized);
+
+        MlKemKeyPair::from(private_key, public_key.into())
+    }
 }
 
 pub(crate) fn encapsulate<
@@ -148,24 +192,45 @@ pub(crate) fn encapsulate<
     let hashed = G(&to_hash);
     let (shared_secret, pseudorandomness) = hashed.split_at(SHARED_SECRET_SIZE);
 
-    let ciphertext = ind_cpa::encrypt::<
-        K,
-        CIPHERTEXT_SIZE,
-        T_AS_NTT_ENCODED_SIZE,
-        C1_SIZE,
-        C2_SIZE,
-        VECTOR_U_COMPRESSION_FACTOR,
-        VECTOR_V_COMPRESSION_FACTOR,
-        VECTOR_U_BLOCK_LEN,
-        ETA1,
-        ETA1_RANDOMNESS_SIZE,
-        ETA2,
-        ETA2_RANDOMNESS_SIZE,
-    >(public_key.as_slice(), randomness, pseudorandomness);
-
-    let mut shared_secret_array = [0u8; constants::SHARED_SECRET_SIZE];
-    shared_secret_array.copy_from_slice(shared_secret);
-    (ciphertext.into(), shared_secret_array)
+    if libcrux_platform::simd256_support() {
+        let ciphertext = ind_cpa::encrypt::<
+            K,
+            CIPHERTEXT_SIZE,
+            T_AS_NTT_ENCODED_SIZE,
+            C1_SIZE,
+            C2_SIZE,
+            VECTOR_U_COMPRESSION_FACTOR,
+            VECTOR_V_COMPRESSION_FACTOR,
+            VECTOR_U_BLOCK_LEN,
+            ETA1,
+            ETA1_RANDOMNESS_SIZE,
+            ETA2,
+            ETA2_RANDOMNESS_SIZE,
+            SIMD256Vector,
+        >(public_key.as_slice(), randomness, pseudorandomness);
+        let mut shared_secret_array = [0u8; constants::SHARED_SECRET_SIZE];
+        shared_secret_array.copy_from_slice(shared_secret);
+        (ciphertext.into(), shared_secret_array)
+    } else {
+        let ciphertext = ind_cpa::encrypt::<
+            K,
+            CIPHERTEXT_SIZE,
+            T_AS_NTT_ENCODED_SIZE,
+            C1_SIZE,
+            C2_SIZE,
+            VECTOR_U_COMPRESSION_FACTOR,
+            VECTOR_V_COMPRESSION_FACTOR,
+            VECTOR_U_BLOCK_LEN,
+            ETA1,
+            ETA1_RANDOMNESS_SIZE,
+            ETA2,
+            ETA2_RANDOMNESS_SIZE,
+            PortableVector,
+        >(public_key.as_slice(), randomness, pseudorandomness);
+        let mut shared_secret_array = [0u8; constants::SHARED_SECRET_SIZE];
+        shared_secret_array.copy_from_slice(shared_secret);
+        (ciphertext.into(), shared_secret_array)
+    }
 }
 
 pub(crate) fn decapsulate<
@@ -193,48 +258,99 @@ pub(crate) fn decapsulate<
     let (ind_cpa_public_key, secret_key) = secret_key.split_at(PUBLIC_KEY_SIZE);
     let (ind_cpa_public_key_hash, implicit_rejection_value) = secret_key.split_at(H_DIGEST_SIZE);
 
-    let decrypted = ind_cpa::decrypt::<
-        K,
-        CIPHERTEXT_SIZE,
-        C1_SIZE,
-        VECTOR_U_COMPRESSION_FACTOR,
-        VECTOR_V_COMPRESSION_FACTOR,
-    >(ind_cpa_secret_key, &ciphertext.value);
+    if libcrux_platform::simd256_support() {
+        let decrypted = ind_cpa::decrypt::<
+            K,
+            CIPHERTEXT_SIZE,
+            C1_SIZE,
+            VECTOR_U_COMPRESSION_FACTOR,
+            VECTOR_V_COMPRESSION_FACTOR,
+            SIMD256Vector,
+        >(ind_cpa_secret_key, &ciphertext.value);
 
-    let mut to_hash: [u8; SHARED_SECRET_SIZE + H_DIGEST_SIZE] = into_padded_array(&decrypted);
-    to_hash[SHARED_SECRET_SIZE..].copy_from_slice(ind_cpa_public_key_hash);
+        let mut to_hash: [u8; SHARED_SECRET_SIZE + H_DIGEST_SIZE] = into_padded_array(&decrypted);
+        to_hash[SHARED_SECRET_SIZE..].copy_from_slice(ind_cpa_public_key_hash);
 
-    let hashed = G(&to_hash);
-    let (shared_secret, pseudorandomness) = hashed.split_at(SHARED_SECRET_SIZE);
+        let hashed = G(&to_hash);
+        let (shared_secret, pseudorandomness) = hashed.split_at(SHARED_SECRET_SIZE);
 
-    let mut to_hash: [u8; IMPLICIT_REJECTION_HASH_INPUT_SIZE] =
-        into_padded_array(&implicit_rejection_value);
-    to_hash[SHARED_SECRET_SIZE..].copy_from_slice(ciphertext.as_ref());
-    let implicit_rejection_shared_secret: [u8; SHARED_SECRET_SIZE] = PRF(&to_hash);
+        let mut to_hash: [u8; IMPLICIT_REJECTION_HASH_INPUT_SIZE] =
+            into_padded_array(&implicit_rejection_value);
+        to_hash[SHARED_SECRET_SIZE..].copy_from_slice(ciphertext.as_ref());
+        let implicit_rejection_shared_secret: [u8; SHARED_SECRET_SIZE] = PRF(&to_hash);
 
-    let expected_ciphertext = ind_cpa::encrypt::<
-        K,
-        CIPHERTEXT_SIZE,
-        T_AS_NTT_ENCODED_SIZE,
-        C1_SIZE,
-        C2_SIZE,
-        VECTOR_U_COMPRESSION_FACTOR,
-        VECTOR_V_COMPRESSION_FACTOR,
-        C1_BLOCK_SIZE,
-        ETA1,
-        ETA1_RANDOMNESS_SIZE,
-        ETA2,
-        ETA2_RANDOMNESS_SIZE,
-    >(ind_cpa_public_key, decrypted, pseudorandomness);
+        let expected_ciphertext = ind_cpa::encrypt::<
+            K,
+            CIPHERTEXT_SIZE,
+            T_AS_NTT_ENCODED_SIZE,
+            C1_SIZE,
+            C2_SIZE,
+            VECTOR_U_COMPRESSION_FACTOR,
+            VECTOR_V_COMPRESSION_FACTOR,
+            C1_BLOCK_SIZE,
+            ETA1,
+            ETA1_RANDOMNESS_SIZE,
+            ETA2,
+            ETA2_RANDOMNESS_SIZE,
+            SIMD256Vector,
+        >(ind_cpa_public_key, decrypted, pseudorandomness);
 
-    let selector = compare_ciphertexts_in_constant_time::<CIPHERTEXT_SIZE>(
-        ciphertext.as_ref(),
-        &expected_ciphertext,
-    );
+        let selector = compare_ciphertexts_in_constant_time::<CIPHERTEXT_SIZE>(
+            ciphertext.as_ref(),
+            &expected_ciphertext,
+        );
 
-    select_shared_secret_in_constant_time(
-        shared_secret,
-        &implicit_rejection_shared_secret,
-        selector,
-    )
+        select_shared_secret_in_constant_time(
+            shared_secret,
+            &implicit_rejection_shared_secret,
+            selector,
+        )
+    } else {
+        let decrypted = ind_cpa::decrypt::<
+            K,
+            CIPHERTEXT_SIZE,
+            C1_SIZE,
+            VECTOR_U_COMPRESSION_FACTOR,
+            VECTOR_V_COMPRESSION_FACTOR,
+            PortableVector,
+        >(ind_cpa_secret_key, &ciphertext.value);
+
+        let mut to_hash: [u8; SHARED_SECRET_SIZE + H_DIGEST_SIZE] = into_padded_array(&decrypted);
+        to_hash[SHARED_SECRET_SIZE..].copy_from_slice(ind_cpa_public_key_hash);
+
+        let hashed = G(&to_hash);
+        let (shared_secret, pseudorandomness) = hashed.split_at(SHARED_SECRET_SIZE);
+
+        let mut to_hash: [u8; IMPLICIT_REJECTION_HASH_INPUT_SIZE] =
+            into_padded_array(&implicit_rejection_value);
+        to_hash[SHARED_SECRET_SIZE..].copy_from_slice(ciphertext.as_ref());
+        let implicit_rejection_shared_secret: [u8; SHARED_SECRET_SIZE] = PRF(&to_hash);
+
+        let expected_ciphertext = ind_cpa::encrypt::<
+            K,
+            CIPHERTEXT_SIZE,
+            T_AS_NTT_ENCODED_SIZE,
+            C1_SIZE,
+            C2_SIZE,
+            VECTOR_U_COMPRESSION_FACTOR,
+            VECTOR_V_COMPRESSION_FACTOR,
+            C1_BLOCK_SIZE,
+            ETA1,
+            ETA1_RANDOMNESS_SIZE,
+            ETA2,
+            ETA2_RANDOMNESS_SIZE,
+            PortableVector,
+        >(ind_cpa_public_key, decrypted, pseudorandomness);
+
+        let selector = compare_ciphertexts_in_constant_time::<CIPHERTEXT_SIZE>(
+            ciphertext.as_ref(),
+            &expected_ciphertext,
+        );
+
+        select_shared_secret_in_constant_time(
+            shared_secret,
+            &implicit_rejection_shared_secret,
+            selector,
+        )
+    }
 }
