+++ libcrux-ml-kem/src/matrix.rs
@@ -1,20 +1,16 @@
+use libcrux_polynomials::Operations;
+
 use crate::{
-    helper::cloop,
-    ntt::invert_ntt_montgomery,
-    polynomial::{
-        add_error_reduce, add_message_error_reduce, add_standard_error_reduce, add_to_ring_element,
-        ntt_multiply, subtract_reduce, PolynomialRingElement,
-    },
-    sampling::sample_from_xof,
+    helper::cloop, ntt::invert_ntt_montgomery, polynomial::PolynomialRingElement, sampling::sample_from_xof
 };
 
 #[inline(always)]
 #[allow(non_snake_case)]
-pub(crate) fn sample_matrix_A<const K: usize>(
+pub(crate) fn sample_matrix_A<const K: usize, Vector: Operations>(
     seed: [u8; 34],
     transpose: bool,
-) -> [[PolynomialRingElement; K]; K] {
-    let mut A_transpose = [[PolynomialRingElement::ZERO(); K]; K];
+) -> [[PolynomialRingElement<Vector>; K]; K] {
+    let mut A_transpose = [[PolynomialRingElement::<Vector>::ZERO(); K]; K];
 
     for i in 0..K {
         let mut seeds = [seed; K];
@@ -42,65 +38,65 @@ pub(crate) fn sample_matrix_A<const K: usize>(
 
 /// Compute v − InverseNTT(sᵀ ◦ NTT(u))
 #[inline(always)]
-pub(crate) fn compute_message<const K: usize>(
-    v: &PolynomialRingElement,
-    secret_as_ntt: &[PolynomialRingElement; K],
-    u_as_ntt: &[PolynomialRingElement; K],
-) -> PolynomialRingElement {
-    let mut result = PolynomialRingElement::ZERO();
+pub(crate) fn compute_message<const K: usize, Vector: Operations>(
+    v: &PolynomialRingElement<Vector>,
+    secret_as_ntt: &[PolynomialRingElement<Vector>; K],
+    u_as_ntt: &[PolynomialRingElement<Vector>; K],
+) -> PolynomialRingElement<Vector> {
+    let mut result = PolynomialRingElement::<Vector>::ZERO();
 
     for i in 0..K {
-        let product = ntt_multiply(&secret_as_ntt[i], &u_as_ntt[i]);
-        result = add_to_ring_element::<K>(result, &product);
+        let product = secret_as_ntt[i].ntt_multiply(&u_as_ntt[i]);
+        result = result.add_to_ring_element::<K>(&product);
     }
 
-    result = invert_ntt_montgomery::<K>(result);
-    result = subtract_reduce(v, result);
+    result = invert_ntt_montgomery::<K ,Vector>(result);
+    result = v.subtract_reduce(result);
 
     result
 }
 
 /// Compute InverseNTT(tᵀ ◦ r̂) + e₂ + message
 #[inline(always)]
-pub(crate) fn compute_ring_element_v<const K: usize>(
-    t_as_ntt: &[PolynomialRingElement; K],
-    r_as_ntt: &[PolynomialRingElement; K],
-    error_2: &PolynomialRingElement,
-    message: &PolynomialRingElement,
-) -> PolynomialRingElement {
-    let mut result = PolynomialRingElement::ZERO();
+pub(crate) fn compute_ring_element_v<const K: usize, Vector: Operations>(
+    t_as_ntt: &[PolynomialRingElement<Vector>; K],
+    r_as_ntt: &[PolynomialRingElement<Vector>; K],
+    error_2: &PolynomialRingElement<Vector>,
+    message: &PolynomialRingElement<Vector>,
+) -> PolynomialRingElement<Vector> {
+    let mut result = PolynomialRingElement::<Vector>::ZERO();
 
     for i in 0..K {
-        let product = ntt_multiply(&t_as_ntt[i], &r_as_ntt[i]);
-        result = add_to_ring_element::<K>(result, &product);
+        let product = t_as_ntt[i].ntt_multiply(&r_as_ntt[i]);
+        result = result.add_to_ring_element::<K>(&product);
     }
 
-    result = invert_ntt_montgomery::<K>(result);
-    result = add_message_error_reduce(error_2, message, result);
+    result = invert_ntt_montgomery::<K, Vector>(result);
+    result = error_2.add_message_error_reduce(message, result);
 
     result
 }
 
 /// Compute u := InvertNTT(Aᵀ ◦ r̂) + e₁
 #[inline(always)]
-pub(crate) fn compute_vector_u<const K: usize>(
-    a_as_ntt: &[[PolynomialRingElement; K]; K],
-    r_as_ntt: &[PolynomialRingElement; K],
-    error_1: &[PolynomialRingElement; K],
-) -> [PolynomialRingElement; K] {
-    let mut result = [PolynomialRingElement::ZERO(); K];
+pub(crate) fn compute_vector_u<const K: usize, Vector: Operations>(
+    a_as_ntt: &[[PolynomialRingElement<Vector>; K]; K],
+    r_as_ntt: &[PolynomialRingElement<Vector>; K],
+    error_1: &[PolynomialRingElement<Vector>; K],
+) -> [PolynomialRingElement<Vector>; K] {
+    let mut result = [PolynomialRingElement::<Vector>::ZERO(); K];
 
     cloop! {
         for (i, row) in a_as_ntt.iter().enumerate() {
             cloop! {
                 for (j, a_element) in row.iter().enumerate() {
-                    let product = ntt_multiply(a_element, &r_as_ntt[j]);
-                    result[i] = add_to_ring_element::<K>(result[i], &product);
+                    let product = a_element.ntt_multiply(&r_as_ntt[j]);
+                    result[i] = result[i].add_to_ring_element::<K>(&product);
                 }
             }
 
-            result[i] = invert_ntt_montgomery::<K>(result[i]);
-            result[i] = add_error_reduce(&error_1[i],result[i]);
+            result[i] = invert_ntt_montgomery::<K, Vector>(result[i]);
+            result[i] = error_1[i].add_error_reduce(result[i]);
         }
     }
 
@@ -110,22 +106,22 @@ pub(crate) fn compute_vector_u<const K: usize>(
 /// Compute Â ◦ ŝ + ê
 #[inline(always)]
 #[allow(non_snake_case)]
-pub(crate) fn compute_As_plus_e<const K: usize>(
-    matrix_A: &[[PolynomialRingElement; K]; K],
-    s_as_ntt: &[PolynomialRingElement; K],
-    error_as_ntt: &[PolynomialRingElement; K],
-) -> [PolynomialRingElement; K] {
-    let mut result = [PolynomialRingElement::ZERO(); K];
+pub(crate) fn compute_As_plus_e<const K: usize, Vector: Operations>(
+    matrix_A: &[[PolynomialRingElement<Vector>; K]; K],
+    s_as_ntt: &[PolynomialRingElement<Vector>; K],
+    error_as_ntt: &[PolynomialRingElement<Vector>; K],
+) -> [PolynomialRingElement<Vector>; K] {
+    let mut result = [PolynomialRingElement::<Vector>::ZERO(); K];
 
     cloop! {
         for (i, row) in matrix_A.iter().enumerate() {
             cloop! {
                 for (j, matrix_element) in row.iter().enumerate() {
-                    let product = ntt_multiply(matrix_element, &s_as_ntt[j]);
-                    result[i] = add_to_ring_element::<K>(result[i], &product);
+                    let product = matrix_element.ntt_multiply(&s_as_ntt[j]);
+                    result[i] = result[i].add_to_ring_element::<K>(&product);
                 }
             }
-            result[i] = add_standard_error_reduce(&error_as_ntt[i], result[i]);
+            result[i] = error_as_ntt[i].add_standard_error_reduce(result[i]);
         }
     }
