+++ libcrux-ml-kem/src/polynomial.rs
@@ -1,130 +1,166 @@
-use libcrux_polynomials_avx2::traits::{GenericOperations, Operations, FIELD_ELEMENTS_IN_VECTOR};
-
-//use crate::hax_utils::hax_debug_assert;
-use crate::arithmetic::*;
-use crate::simd;
+use libcrux_polynomials::{
+    FieldElementTimesMontgomeryR, GenericOperations, Operations, FIELD_ELEMENTS_IN_VECTOR,
+};
 
 pub(crate) const VECTORS_IN_RING_ELEMENT: usize =
     super::constants::COEFFICIENTS_IN_RING_ELEMENT / FIELD_ELEMENTS_IN_VECTOR;
 
 #[derive(Clone, Copy)]
-pub struct PolynomialRingElement {
-    pub(crate) coefficients: [simd::Vector; VECTORS_IN_RING_ELEMENT],
+pub(crate) struct PolynomialRingElement<Vector: Operations> {
+    pub(crate) coefficients: [Vector; VECTORS_IN_RING_ELEMENT],
 }
 
-impl PolynomialRingElement {
+impl<Vector: Operations> PolynomialRingElement<Vector> {
     #[allow(non_snake_case)]
     pub(crate) fn ZERO() -> Self {
         Self {
             // FIXME:  The THIR body of item DefId(0:415 ~ libcrux_ml_kem[9000]::polynomial::{impl#0}::ZERO::{constant#0}) was stolen.
-            coefficients: [simd::Vector::ZERO(); 32],
+            coefficients: [Vector::ZERO(); 32],
         }
     }
-}
 
-#[inline(always)]
-pub(crate) fn from_i32_array(a: [i32; 256]) -> PolynomialRingElement {
-    let mut result = PolynomialRingElement::ZERO();
-    for i in 0..VECTORS_IN_RING_ELEMENT {
-        result.coefficients[i] = simd::Vector::from_i32_array(
-            a[i * FIELD_ELEMENTS_IN_VECTOR..(i + 1) * FIELD_ELEMENTS_IN_VECTOR]
-                .try_into()
-                .unwrap(),
-        );
+    #[inline(always)]
+    pub(crate) fn from_i32_array(a: [i32; 256]) -> Self {
+        let mut result = PolynomialRingElement::ZERO();
+        for i in 0..VECTORS_IN_RING_ELEMENT {
+            result.coefficients[i] = Vector::from_i32_array(
+                a[i * FIELD_ELEMENTS_IN_VECTOR..(i + 1) * FIELD_ELEMENTS_IN_VECTOR]
+                    .try_into()
+                    .unwrap(),
+            );
+        }
+        result
     }
-    result
-}
 
-/// Given two polynomial ring elements `lhs` and `rhs`, compute the pointwise
-/// sum of their constituent coefficients.
-#[inline(always)]
-pub(crate) fn add_to_ring_element<const K: usize>(
-    mut lhs: PolynomialRingElement,
-    rhs: &PolynomialRingElement,
-) -> PolynomialRingElement {
-    for i in 0..lhs.coefficients.len() {
-        lhs.coefficients[i] = simd::Vector::add(lhs.coefficients[i], &rhs.coefficients[i]);
+    /// Given two polynomial ring elements `lhs` and `rhs`, compute the pointwise
+    /// sum of their constituent coefficients.
+    #[inline(always)]
+    pub(crate) fn add_to_ring_element<const K: usize>(mut self, rhs: &Self) -> Self {
+        for i in 0..self.coefficients.len() {
+            self.coefficients[i] = Vector::add(self.coefficients[i], &rhs.coefficients[i]);
+        }
+        self
     }
-    lhs
-}
 
-#[inline(always)]
-pub(crate) fn poly_barrett_reduce(mut a: PolynomialRingElement) -> PolynomialRingElement {
-    for i in 0..VECTORS_IN_RING_ELEMENT {
-        a.coefficients[i] = simd::Vector::barrett_reduce(a.coefficients[i]);
+    #[inline(always)]
+    pub fn poly_barrett_reduce(mut self) -> Self {
+        for i in 0..VECTORS_IN_RING_ELEMENT {
+            self.coefficients[i] = Vector::barrett_reduce(self.coefficients[i]);
+        }
+        self
     }
-    a
-}
 
-#[inline(always)]
-pub(crate) fn subtract_reduce(
-    a: &PolynomialRingElement,
-    mut b: PolynomialRingElement,
-) -> PolynomialRingElement {
-    for i in 0..VECTORS_IN_RING_ELEMENT {
-        let coefficient_normal_form = simd::Vector::montgomery_reduce(
-            simd::Vector::multiply_by_constant(b.coefficients[i], 1441),
-        );
-        b.coefficients[i] = simd::Vector::barrett_reduce(simd::Vector::sub(
-            a.coefficients[i],
-            &coefficient_normal_form,
-        ));
+    #[inline(always)]
+    pub(crate) fn subtract_reduce(&self, mut b: Self) -> Self {
+        for i in 0..VECTORS_IN_RING_ELEMENT {
+            let coefficient_normal_form =
+                Vector::montgomery_reduce(Vector::multiply_by_constant(b.coefficients[i], 1441));
+            b.coefficients[i] =
+                Vector::barrett_reduce(Vector::sub(self.coefficients[i], &coefficient_normal_form));
+        }
+        b
     }
-    b
-}
 
-#[inline(always)]
-pub(crate) fn add_message_error_reduce(
-    err: &PolynomialRingElement,
-    message: &PolynomialRingElement,
-    mut result: PolynomialRingElement,
-) -> PolynomialRingElement {
-    for i in 0..VECTORS_IN_RING_ELEMENT {
-        let coefficient_normal_form = simd::Vector::montgomery_reduce(
-            simd::Vector::multiply_by_constant(result.coefficients[i], 1441),
-        );
-        result.coefficients[i] = simd::Vector::barrett_reduce(simd::Vector::add(
-            coefficient_normal_form,
-            &simd::Vector::add(err.coefficients[i], &message.coefficients[i]),
-        ));
+    #[inline(always)]
+    pub(crate) fn add_message_error_reduce(&self, message: &Self, mut result: Self) -> Self {
+        for i in 0..VECTORS_IN_RING_ELEMENT {
+            let coefficient_normal_form = Vector::montgomery_reduce(Vector::multiply_by_constant(
+                result.coefficients[i],
+                1441,
+            ));
+            result.coefficients[i] = Vector::barrett_reduce(Vector::add(
+                coefficient_normal_form,
+                &Vector::add(self.coefficients[i], &message.coefficients[i]),
+            ));
+        }
+        result
     }
-    result
-}
 
-#[inline(always)]
-pub(crate) fn add_error_reduce(
-    err: &PolynomialRingElement,
-    mut result: PolynomialRingElement,
-) -> PolynomialRingElement {
-    for j in 0..VECTORS_IN_RING_ELEMENT {
-        let coefficient_normal_form = simd::Vector::montgomery_reduce(
-            simd::Vector::multiply_by_constant(result.coefficients[j], 1441),
-        );
+    #[inline(always)]
+    pub(crate) fn add_error_reduce(&self, mut result: Self) -> Self {
+        for j in 0..VECTORS_IN_RING_ELEMENT {
+            let coefficient_normal_form = Vector::montgomery_reduce(Vector::multiply_by_constant(
+                result.coefficients[j],
+                1441,
+            ));
 
-        result.coefficients[j] = simd::Vector::barrett_reduce(simd::Vector::add(
-            coefficient_normal_form,
-            &err.coefficients[j],
-        ));
+            result.coefficients[j] =
+                Vector::barrett_reduce(Vector::add(coefficient_normal_form, &self.coefficients[j]));
+        }
+        result
     }
-    result
-}
 
-#[inline(always)]
-pub(crate) fn add_standard_error_reduce(
-    err: &PolynomialRingElement,
-    mut result: PolynomialRingElement,
-) -> PolynomialRingElement {
-    for j in 0..VECTORS_IN_RING_ELEMENT {
-        // The coefficients are of the form aR^{-1} mod q, which means
-        // calling to_montgomery_domain() on them should return a mod q.
-        let coefficient_normal_form = simd::Vector::to_standard_domain(result.coefficients[j]);
+    #[inline(always)]
+    pub(crate) fn add_standard_error_reduce(&self, mut result: Self) -> Self {
+        for j in 0..VECTORS_IN_RING_ELEMENT {
+            // The coefficients are of the form aR^{-1} mod q, which means
+            // calling to_montgomery_domain() on them should return a mod q.
+            let coefficient_normal_form = Vector::to_standard_domain(result.coefficients[j]);
 
-        result.coefficients[j] = simd::Vector::barrett_reduce(simd::Vector::add(
-            coefficient_normal_form,
-            &err.coefficients[j],
-        ));
+            result.coefficients[j] =
+                Vector::barrett_reduce(Vector::add(coefficient_normal_form, &self.coefficients[j]));
+        }
+        result
     }
-    result
+
+    /// Given two `KyberPolynomialRingElement`s in their NTT representations,
+    /// compute their product. Given two polynomials in the NTT domain `f^` and `ĵ`,
+    /// the `iᵗʰ` coefficient of the product `k̂` is determined by the calculation:
+    ///
+    /// ```plaintext
+    /// ĥ[2·i] + ĥ[2·i + 1]X = (f^[2·i] + f^[2·i + 1]X)·(ĝ[2·i] + ĝ[2·i + 1]X) mod (X² - ζ^(2·BitRev₇(i) + 1))
+    /// ```
+    ///
+    /// This function almost implements <strong>Algorithm 10</strong> of the
+    /// NIST FIPS 203 standard, which is reproduced below:
+    ///
+    /// ```plaintext
+    /// Input: Two arrays fˆ ∈ ℤ₂₅₆ and ĝ ∈ ℤ₂₅₆.
+    /// Output: An array ĥ ∈ ℤq.
+    ///
+    /// for(i ← 0; i < 128; i++)
+    ///     (ĥ[2i], ĥ[2i+1]) ← BaseCaseMultiply(fˆ[2i], fˆ[2i+1], ĝ[2i], ĝ[2i+1], ζ^(2·BitRev₇(i) + 1))
+    /// end for
+    /// return ĥ
+    /// ```
+    /// We say "almost" because the coefficients of the ring element output by
+    /// this function are in the Montgomery domain.
+    ///
+    /// The NIST FIPS 203 standard can be found at
+    /// <https://csrc.nist.gov/pubs/fips/203/ipd>.
+    // TODO: Remove or replace with something that works and is useful for the proof.
+    // #[cfg_attr(hax, hax_lib::requires(
+    //     hax_lib::forall(|i:usize|
+    //         hax_lib::implies(i < COEFFICIENTS_IN_RING_ELEMENT, ||
+    //             (lhs.coefficients[i] >= 0 && lhs.coefficients[i] < 4096) &&
+    //             (rhs.coefficients[i].abs() <= FIELD_MODULUS)
+
+    // ))))]
+    // #[cfg_attr(hax, hax_lib::ensures(|result|
+    //     hax_lib::forall(|i:usize|
+    //         hax_lib::implies(i < result.coefficients.len(), ||
+    //                 result.coefficients[i].abs() <= FIELD_MODULUS
+    // ))))]
+    #[inline(always)]
+    pub(crate) fn ntt_multiply(&self, rhs: &Self) -> Self {
+        // hax_debug_debug_assert!(lhs
+        //     .coefficients
+        //     .into_iter()
+        //     .all(|coefficient| coefficient >= 0 && coefficient < 4096));
+
+        let mut out = PolynomialRingElement::ZERO();
+
+        for i in 0..VECTORS_IN_RING_ELEMENT {
+            out.coefficients[i] = Vector::ntt_multiply(
+                &self.coefficients[i],
+                &rhs.coefficients[i],
+                ZETAS_TIMES_MONTGOMERY_R[64 + 2 * i],
+                ZETAS_TIMES_MONTGOMERY_R[64 + 2 * i + 1],
+            );
+        }
+
+        out
+    }
 }
 
 const ZETAS_TIMES_MONTGOMERY_R: [FieldElementTimesMontgomeryR; 128] = [
@@ -142,15 +178,15 @@ const ZETAS_TIMES_MONTGOMERY_R: [FieldElementTimesMontgomeryR; 128] = [
 /// resulting coefficients are in the normal domain since the zetas have been
 /// multiplied by MONTGOMERY_R.
 #[inline(always)]
-pub(crate) fn ntt_at_layer_1(
+pub(crate) fn ntt_at_layer_1<Vector: Operations>(
     zeta_i: &mut usize,
-    mut re: PolynomialRingElement,
+    mut re: PolynomialRingElement<Vector>,
     _layer: usize,
     _initial_coefficient_bound: usize,
-) -> PolynomialRingElement {
+) -> PolynomialRingElement<Vector> {
     *zeta_i += 1;
     for round in 0..32 {
-        re.coefficients[round] = simd::Vector::ntt_layer_1_step(
+        re.coefficients[round] = Vector::ntt_layer_1_step(
             re.coefficients[round],
             ZETAS_TIMES_MONTGOMERY_R[*zeta_i],
             ZETAS_TIMES_MONTGOMERY_R[*zeta_i + 1],
@@ -162,41 +198,39 @@ pub(crate) fn ntt_at_layer_1(
 }
 
 #[inline(always)]
-pub(crate) fn ntt_at_layer_2(
+pub(crate) fn ntt_at_layer_2<Vector: Operations>(
     zeta_i: &mut usize,
-    mut re: PolynomialRingElement,
+    mut re: PolynomialRingElement<Vector>,
     _layer: usize,
     _initial_coefficient_bound: usize,
-) -> PolynomialRingElement {
+) -> PolynomialRingElement<Vector> {
     for round in 0..32 {
         *zeta_i += 1;
-        re.coefficients[round] = simd::Vector::ntt_layer_2_step(
-            re.coefficients[round],
-            ZETAS_TIMES_MONTGOMERY_R[*zeta_i],
-        );
+        re.coefficients[round] =
+            Vector::ntt_layer_2_step(re.coefficients[round], ZETAS_TIMES_MONTGOMERY_R[*zeta_i]);
     }
     re
 }
 
 #[inline(always)]
-pub(crate) fn ntt_layer_int_vec_step(
-    mut a: simd::Vector,
-    mut b: simd::Vector,
+pub(crate) fn ntt_layer_int_vec_step<Vector: Operations>(
+    mut a: Vector,
+    mut b: Vector,
     zeta_r: i32,
-) -> (simd::Vector, simd::Vector) {
-    let t = simd::Vector::montgomery_multiply_fe_by_fer(b, zeta_r);
-    b = simd::Vector::sub(a, &t);
-    a = simd::Vector::add(a, &t);
+) -> (Vector, Vector) {
+    let t = Vector::montgomery_multiply_fe_by_fer(b, zeta_r);
+    b = Vector::sub(a, &t);
+    a = Vector::add(a, &t);
     (a, b)
 }
 
 #[inline(always)]
-pub(crate) fn ntt_at_layer_3_plus(
+pub(crate) fn ntt_at_layer_3_plus<Vector: Operations>(
     zeta_i: &mut usize,
-    mut re: PolynomialRingElement,
+    mut re: PolynomialRingElement<Vector>,
     layer: usize,
     _initial_coefficient_bound: usize,
-) -> PolynomialRingElement {
+) -> PolynomialRingElement<Vector> {
     debug_assert!(layer >= 3);
     let step = 1 << layer;
 
@@ -221,18 +255,20 @@ pub(crate) fn ntt_at_layer_3_plus(
 }
 
 #[inline(always)]
-pub(crate) fn ntt_layer_7_int_vec_step(
-    mut a: simd::Vector,
-    mut b: simd::Vector,
-) -> (simd::Vector, simd::Vector) {
-    let t = simd::Vector::multiply_by_constant(b, -1600);
-    b = simd::Vector::sub(a, &t);
-    a = simd::Vector::add(a, &t);
+pub(crate) fn ntt_layer_7_int_vec_step<Vector: Operations>(
+    mut a: Vector,
+    mut b: Vector,
+) -> (Vector, Vector) {
+    let t = Vector::multiply_by_constant(b, -1600);
+    b = Vector::sub(a, &t);
+    a = Vector::add(a, &t);
     (a, b)
 }
 
 #[inline(always)]
-pub(crate) fn ntt_at_layer_7(mut re: PolynomialRingElement) -> PolynomialRingElement {
+pub(crate) fn ntt_at_layer_7<Vector: Operations>(
+    mut re: PolynomialRingElement<Vector>,
+) -> PolynomialRingElement<Vector> {
     let step = VECTORS_IN_RING_ELEMENT / 2;
     for j in 0..step {
         let (x, y) = ntt_layer_7_int_vec_step(re.coefficients[j], re.coefficients[j + step]);
@@ -243,14 +279,14 @@ pub(crate) fn ntt_at_layer_7(mut re: PolynomialRingElement) -> PolynomialRingEle
 }
 
 #[inline(always)]
-pub(crate) fn invert_ntt_at_layer_1(
+pub(crate) fn invert_ntt_at_layer_1<Vector: Operations>(
     zeta_i: &mut usize,
-    mut re: PolynomialRingElement,
+    mut re: PolynomialRingElement<Vector>,
     _layer: usize,
-) -> PolynomialRingElement {
+) -> PolynomialRingElement<Vector> {
     *zeta_i -= 1;
     for round in 0..32 {
-        re.coefficients[round] = simd::Vector::inv_ntt_layer_1_step(
+        re.coefficients[round] = Vector::inv_ntt_layer_1_step(
             re.coefficients[round],
             ZETAS_TIMES_MONTGOMERY_R[*zeta_i],
             ZETAS_TIMES_MONTGOMERY_R[*zeta_i - 1],
@@ -262,39 +298,37 @@ pub(crate) fn invert_ntt_at_layer_1(
 }
 
 #[inline(always)]
-pub(crate) fn invert_ntt_at_layer_2(
+pub(crate) fn invert_ntt_at_layer_2<Vector: Operations>(
     zeta_i: &mut usize,
-    mut re: PolynomialRingElement,
+    mut re: PolynomialRingElement<Vector>,
     _layer: usize,
-) -> PolynomialRingElement {
+) -> PolynomialRingElement<Vector> {
     for round in 0..32 {
         *zeta_i -= 1;
-        re.coefficients[round] = simd::Vector::inv_ntt_layer_2_step(
-            re.coefficients[round],
-            ZETAS_TIMES_MONTGOMERY_R[*zeta_i],
-        );
+        re.coefficients[round] =
+            Vector::inv_ntt_layer_2_step(re.coefficients[round], ZETAS_TIMES_MONTGOMERY_R[*zeta_i]);
     }
     re
 }
 
 #[inline(always)]
-pub(crate) fn inv_ntt_layer_int_vec_step(
-    mut a: simd::Vector,
-    mut b: simd::Vector,
+pub(crate) fn inv_ntt_layer_int_vec_step<Vector: Operations>(
+    mut a: Vector,
+    mut b: Vector,
     zeta_r: i32,
-) -> (simd::Vector, simd::Vector) {
-    let a_minus_b = simd::Vector::sub(b, &a);
-    a = simd::Vector::add(a, &b);
-    b = simd::Vector::montgomery_multiply_fe_by_fer(a_minus_b, zeta_r);
+) -> (Vector, Vector) {
+    let a_minus_b = Vector::sub(b, &a);
+    a = Vector::add(a, &b);
+    b = Vector::montgomery_multiply_fe_by_fer(a_minus_b, zeta_r);
     (a, b)
 }
 
 #[inline(always)]
-pub(crate) fn invert_ntt_at_layer_3_plus(
+pub(crate) fn invert_ntt_at_layer_3_plus<Vector: Operations>(
     zeta_i: &mut usize,
-    mut re: PolynomialRingElement,
+    mut re: PolynomialRingElement<Vector>,
     layer: usize,
-) -> PolynomialRingElement {
+) -> PolynomialRingElement<Vector> {
     let step = 1 << layer;
 
     for round in 0..(128 >> layer) {
@@ -315,66 +349,4 @@ pub(crate) fn invert_ntt_at_layer_3_plus(
         }
     }
     re
-}
-
-/// Given two `KyberPolynomialRingElement`s in their NTT representations,
-/// compute their product. Given two polynomials in the NTT domain `f^` and `ĵ`,
-/// the `iᵗʰ` coefficient of the product `k̂` is determined by the calculation:
-///
-/// ```plaintext
-/// ĥ[2·i] + ĥ[2·i + 1]X = (f^[2·i] + f^[2·i + 1]X)·(ĝ[2·i] + ĝ[2·i + 1]X) mod (X² - ζ^(2·BitRev₇(i) + 1))
-/// ```
-///
-/// This function almost implements <strong>Algorithm 10</strong> of the
-/// NIST FIPS 203 standard, which is reproduced below:
-///
-/// ```plaintext
-/// Input: Two arrays fˆ ∈ ℤ₂₅₆ and ĝ ∈ ℤ₂₅₆.
-/// Output: An array ĥ ∈ ℤq.
-///
-/// for(i ← 0; i < 128; i++)
-///     (ĥ[2i], ĥ[2i+1]) ← BaseCaseMultiply(fˆ[2i], fˆ[2i+1], ĝ[2i], ĝ[2i+1], ζ^(2·BitRev₇(i) + 1))
-/// end for
-/// return ĥ
-/// ```
-/// We say "almost" because the coefficients of the ring element output by
-/// this function are in the Montgomery domain.
-///
-/// The NIST FIPS 203 standard can be found at
-/// <https://csrc.nist.gov/pubs/fips/203/ipd>.
-// TODO: Remove or replace with something that works and is useful for the proof.
-// #[cfg_attr(hax, hax_lib::requires(
-//     hax_lib::forall(|i:usize|
-//         hax_lib::implies(i < COEFFICIENTS_IN_RING_ELEMENT, ||
-//             (lhs.coefficients[i] >= 0 && lhs.coefficients[i] < 4096) &&
-//             (rhs.coefficients[i].abs() <= FIELD_MODULUS)
-
-// ))))]
-// #[cfg_attr(hax, hax_lib::ensures(|result|
-//     hax_lib::forall(|i:usize|
-//         hax_lib::implies(i < result.coefficients.len(), ||
-//                 result.coefficients[i].abs() <= FIELD_MODULUS
-// ))))]
-#[inline(always)]
-pub(crate) fn ntt_multiply(
-    lhs: &PolynomialRingElement,
-    rhs: &PolynomialRingElement,
-) -> PolynomialRingElement {
-    // hax_debug_debug_assert!(lhs
-    //     .coefficients
-    //     .into_iter()
-    //     .all(|coefficient| coefficient >= 0 && coefficient < 4096));
-
-    let mut out = PolynomialRingElement::ZERO();
-
-    for i in 0..VECTORS_IN_RING_ELEMENT {
-        out.coefficients[i] = simd::Vector::ntt_multiply(
-            &lhs.coefficients[i],
-            &rhs.coefficients[i],
-            ZETAS_TIMES_MONTGOMERY_R[64 + 2 * i],
-            ZETAS_TIMES_MONTGOMERY_R[64 + 2 * i + 1],
-        );
-    }
-
-    out
 }
