+++ libcrux-ml-kem/src/serialize.rs
@@ -3,45 +3,44 @@ use crate::{
     hax_utils::hax_debug_assert,
     helper::cloop,
     polynomial::{PolynomialRingElement, VECTORS_IN_RING_ELEMENT},
-    simd,
 };
-use libcrux_polynomials_avx2::traits::{GenericOperations, Operations};
+use libcrux_polynomials::{GenericOperations, Operations};
 
 #[cfg(hax)]
 use super::constants::COEFFICIENTS_IN_RING_ELEMENT;
 
 #[inline(always)]
-pub(super) fn compress_then_serialize_message(
-    re: PolynomialRingElement,
+pub(super) fn compress_then_serialize_message<Vector: Operations>(
+    re: PolynomialRingElement<Vector>,
 ) -> [u8; SHARED_SECRET_SIZE] {
     let mut serialized = [0u8; SHARED_SECRET_SIZE];
     for i in 0..32 {
-        let coefficient = simd::Vector::to_unsigned_representative(re.coefficients[i]);
-        let coefficient_compressed = simd::Vector::compress_1(coefficient);
-        serialized[i] = simd::Vector::serialize_1(coefficient_compressed);
+        let coefficient = Vector::to_unsigned_representative(re.coefficients[i]);
+        let coefficient_compressed = Vector::compress_1(coefficient);
+        serialized[i] = Vector::serialize_1(coefficient_compressed);
     }
     serialized
 }
 #[inline(always)]
-pub(super) fn deserialize_then_decompress_message(
+pub(super) fn deserialize_then_decompress_message<Vector: Operations>(
     serialized: [u8; SHARED_SECRET_SIZE],
-) -> PolynomialRingElement {
-    let mut re = PolynomialRingElement::ZERO();
+) -> PolynomialRingElement<Vector> {
+    let mut re = PolynomialRingElement::<Vector>::ZERO();
     for i in 0..32 {
-        let coefficient_compressed = simd::Vector::deserialize_1(serialized[i]);
-        re.coefficients[i] = simd::Vector::decompress_1(coefficient_compressed);
+        let coefficient_compressed = Vector::deserialize_1(serialized[i]);
+        re.coefficients[i] = Vector::decompress_1(coefficient_compressed);
     }
     re
 }
 
 #[inline(always)]
-pub(super) fn serialize_uncompressed_ring_element(
-    re: PolynomialRingElement,
+pub(super) fn serialize_uncompressed_ring_element<Vector: Operations>(
+    re: PolynomialRingElement<Vector>,
 ) -> [u8; BYTES_PER_RING_ELEMENT] {
     let mut serialized = [0u8; BYTES_PER_RING_ELEMENT];
     for i in 0..VECTORS_IN_RING_ELEMENT {
-        let coefficient = simd::Vector::to_unsigned_representative(re.coefficients[i]);
-        let bytes = simd::Vector::serialize_12(coefficient);
+        let coefficient = Vector::to_unsigned_representative(re.coefficients[i]);
+        let bytes = Vector::serialize_12(coefficient);
         serialized[12 * i] = bytes[0];
         serialized[12 * i + 1] = bytes[1];
         serialized[12 * i + 2] = bytes[2];
@@ -59,14 +58,16 @@ pub(super) fn serialize_uncompressed_ring_element(
 }
 
 #[inline(always)]
-pub(super) fn deserialize_to_uncompressed_ring_element(serialized: &[u8]) -> PolynomialRingElement {
+pub(super) fn deserialize_to_uncompressed_ring_element<Vector: Operations>(
+    serialized: &[u8],
+) -> PolynomialRingElement<Vector> {
     hax_debug_assert!(serialized.len() == BYTES_PER_RING_ELEMENT);
 
-    let mut re = PolynomialRingElement::ZERO();
+    let mut re = PolynomialRingElement::<Vector>::ZERO();
 
     cloop! {
         for (i, bytes) in serialized.chunks_exact(12).enumerate() {
-            re.coefficients[i] = simd::Vector::deserialize_12(&bytes);
+            re.coefficients[i] = Vector::deserialize_12(&bytes);
         }
     }
 
@@ -77,15 +78,17 @@ pub(super) fn deserialize_to_uncompressed_ring_element(serialized: &[u8]) -> Pol
 ///
 /// This MUST NOT be used with secret inputs, like its caller `deserialize_ring_elements_reduced`.
 #[inline(always)]
-fn deserialize_to_reduced_ring_element(serialized: &[u8]) -> PolynomialRingElement {
+fn deserialize_to_reduced_ring_element<Vector: Operations>(
+    serialized: &[u8],
+) -> PolynomialRingElement<Vector> {
     hax_debug_assert!(serialized.len() == BYTES_PER_RING_ELEMENT);
 
-    let mut re = PolynomialRingElement::ZERO();
+    let mut re = PolynomialRingElement::<Vector>::ZERO();
 
     cloop! {
         for (i, bytes) in serialized.chunks_exact(12).enumerate() {
-            let coefficient = simd::Vector::deserialize_12(&bytes);
-            re.coefficients[i] = simd::Vector::cond_subtract_3329(coefficient);
+            let coefficient = Vector::deserialize_12(&bytes);
+            re.coefficients[i] = Vector::cond_subtract_3329(coefficient);
         }
     }
     re
@@ -96,10 +99,14 @@ fn deserialize_to_reduced_ring_element(serialized: &[u8]) -> PolynomialRingEleme
 ///
 /// This function MUST NOT be used on secret inputs.
 #[inline(always)]
-pub(super) fn deserialize_ring_elements_reduced<const PUBLIC_KEY_SIZE: usize, const K: usize>(
+pub(super) fn deserialize_ring_elements_reduced<
+    const PUBLIC_KEY_SIZE: usize,
+    const K: usize,
+    Vector: Operations,
+>(
     public_key: &[u8],
-) -> [PolynomialRingElement; K] {
-    let mut deserialized_pk = [PolynomialRingElement::ZERO(); K];
+) -> [PolynomialRingElement<Vector>; K] {
+    let mut deserialized_pk = [PolynomialRingElement::<Vector>::ZERO(); K];
     cloop! {
         for (i, ring_element) in public_key
             .chunks_exact(BYTES_PER_RING_ELEMENT)
@@ -112,13 +119,14 @@ pub(super) fn deserialize_ring_elements_reduced<const PUBLIC_KEY_SIZE: usize, co
 }
 
 #[inline(always)]
-fn compress_then_serialize_10<const OUT_LEN: usize>(re: PolynomialRingElement) -> [u8; OUT_LEN] {
+fn compress_then_serialize_10<const OUT_LEN: usize, Vector: Operations>(
+    re: PolynomialRingElement<Vector>,
+) -> [u8; OUT_LEN] {
     let mut serialized = [0u8; OUT_LEN];
     for i in 0..VECTORS_IN_RING_ELEMENT {
-        let coefficient = simd::Vector::compress::<10>(simd::Vector::to_unsigned_representative(
-            re.coefficients[i],
-        ));
-        let bytes = simd::Vector::serialize_10(coefficient);
+        let coefficient =
+            Vector::compress::<10>(Vector::to_unsigned_representative(re.coefficients[i]));
+        let bytes = Vector::serialize_10(coefficient);
         serialized[10 * i] = bytes[0];
         serialized[10 * i + 1] = bytes[1];
         serialized[10 * i + 2] = bytes[2];
@@ -134,13 +142,14 @@ fn compress_then_serialize_10<const OUT_LEN: usize>(re: PolynomialRingElement) -
 }
 
 #[inline(always)]
-fn compress_then_serialize_11<const OUT_LEN: usize>(re: PolynomialRingElement) -> [u8; OUT_LEN] {
+fn compress_then_serialize_11<const OUT_LEN: usize, Vector: Operations>(
+    re: PolynomialRingElement<Vector>,
+) -> [u8; OUT_LEN] {
     let mut serialized = [0u8; OUT_LEN];
     for i in 0..VECTORS_IN_RING_ELEMENT {
-        let coefficient = simd::Vector::compress::<11>(simd::Vector::to_unsigned_representative(
-            re.coefficients[i],
-        ));
-        let bytes = simd::Vector::serialize_11(coefficient);
+        let coefficient =
+            Vector::compress::<11>(Vector::to_unsigned_representative(re.coefficients[i]));
+        let bytes = Vector::serialize_11(coefficient);
         serialized[11 * i] = bytes[0];
         serialized[11 * i + 1] = bytes[1];
         serialized[11 * i + 2] = bytes[2];
@@ -160,8 +169,9 @@ fn compress_then_serialize_11<const OUT_LEN: usize>(re: PolynomialRingElement) -
 pub(super) fn compress_then_serialize_ring_element_u<
     const COMPRESSION_FACTOR: usize,
     const OUT_LEN: usize,
+    Vector: Operations,
 >(
-    re: PolynomialRingElement,
+    re: PolynomialRingElement<Vector>,
 ) -> [u8; OUT_LEN] {
     hax_debug_assert!((COEFFICIENTS_IN_RING_ELEMENT * COMPRESSION_FACTOR) / 8 == OUT_LEN);
 
@@ -173,13 +183,14 @@ pub(super) fn compress_then_serialize_ring_element_u<
 }
 
 #[inline(always)]
-fn compress_then_serialize_4<const OUT_LEN: usize>(re: PolynomialRingElement) -> [u8; OUT_LEN] {
+fn compress_then_serialize_4<const OUT_LEN: usize, Vector: Operations>(
+    re: PolynomialRingElement<Vector>,
+) -> [u8; OUT_LEN] {
     let mut serialized = [0u8; OUT_LEN];
     for i in 0..VECTORS_IN_RING_ELEMENT {
-        let coefficient = simd::Vector::compress::<4>(simd::Vector::to_unsigned_representative(
-            re.coefficients[i],
-        ));
-        let bytes = simd::Vector::serialize_4(coefficient);
+        let coefficient =
+            Vector::compress::<4>(Vector::to_unsigned_representative(re.coefficients[i]));
+        let bytes = Vector::serialize_4(coefficient);
         serialized[4 * i] = bytes[0];
         serialized[4 * i + 1] = bytes[1];
         serialized[4 * i + 2] = bytes[2];
@@ -189,14 +200,15 @@ fn compress_then_serialize_4<const OUT_LEN: usize>(re: PolynomialRingElement) ->
 }
 
 #[inline(always)]
-fn compress_then_serialize_5<const OUT_LEN: usize>(re: PolynomialRingElement) -> [u8; OUT_LEN] {
+fn compress_then_serialize_5<const OUT_LEN: usize, Vector: Operations>(
+    re: PolynomialRingElement<Vector>,
+) -> [u8; OUT_LEN] {
     let mut serialized = [0u8; OUT_LEN];
 
     for i in 0..VECTORS_IN_RING_ELEMENT {
-        let coefficients = simd::Vector::compress::<5>(simd::Vector::to_unsigned_representative(
-            re.coefficients[i],
-        ));
-        let bytes5 = simd::Vector::serialize_5(coefficients);
+        let coefficients =
+            Vector::compress::<5>(Vector::to_unsigned_representative(re.coefficients[i]));
+        let bytes5 = Vector::serialize_5(coefficients);
         serialized[5 * i] = bytes5[0];
         serialized[5 * i + 1] = bytes5[1];
         serialized[5 * i + 2] = bytes5[2];
@@ -210,8 +222,9 @@ fn compress_then_serialize_5<const OUT_LEN: usize>(re: PolynomialRingElement) ->
 pub(super) fn compress_then_serialize_ring_element_v<
     const COMPRESSION_FACTOR: usize,
     const OUT_LEN: usize,
+    Vector: Operations,
 >(
-    re: PolynomialRingElement,
+    re: PolynomialRingElement<Vector>,
 ) -> [u8; OUT_LEN] {
     hax_debug_assert!((COEFFICIENTS_IN_RING_ELEMENT * COMPRESSION_FACTOR) / 8 == OUT_LEN);
 
@@ -223,30 +236,34 @@ pub(super) fn compress_then_serialize_ring_element_v<
 }
 
 #[inline(always)]
-fn deserialize_then_decompress_10(serialized: &[u8]) -> PolynomialRingElement {
+fn deserialize_then_decompress_10<Vector: Operations>(
+    serialized: &[u8],
+) -> PolynomialRingElement<Vector> {
     hax_debug_assert!(serialized.len() == (COEFFICIENTS_IN_RING_ELEMENT * 10) / 8);
 
-    let mut re = PolynomialRingElement::ZERO();
+    let mut re = PolynomialRingElement::<Vector>::ZERO();
 
     cloop! {
         for (i, bytes) in serialized.chunks_exact(10).enumerate() {
-            let coefficient = simd::Vector::deserialize_10(&bytes);
-            re.coefficients[i] = simd::Vector::decompress::<10>(coefficient);
+            let coefficient = Vector::deserialize_10(&bytes);
+            re.coefficients[i] = Vector::decompress::<10>(coefficient);
         }
     }
     re
 }
 
 #[inline(always)]
-fn deserialize_then_decompress_11(serialized: &[u8]) -> PolynomialRingElement {
+fn deserialize_then_decompress_11<Vector: Operations>(
+    serialized: &[u8],
+) -> PolynomialRingElement<Vector> {
     hax_debug_assert!(serialized.len() == (COEFFICIENTS_IN_RING_ELEMENT * 11) / 8);
 
-    let mut re = PolynomialRingElement::ZERO();
+    let mut re = PolynomialRingElement::<Vector>::ZERO();
 
     cloop! {
         for (i, bytes) in serialized.chunks_exact(11).enumerate() {
-            let coefficient = simd::Vector::deserialize_11(&bytes);
-            re.coefficients[i] = simd::Vector::decompress::<11>(coefficient);
+            let coefficient = Vector::deserialize_11(&bytes);
+            re.coefficients[i] = Vector::decompress::<11>(coefficient);
         }
     }
 
@@ -254,9 +271,12 @@ fn deserialize_then_decompress_11(serialized: &[u8]) -> PolynomialRingElement {
 }
 
 #[inline(always)]
-pub(super) fn deserialize_then_decompress_ring_element_u<const COMPRESSION_FACTOR: usize>(
+pub(super) fn deserialize_then_decompress_ring_element_u<
+    const COMPRESSION_FACTOR: usize,
+    Vector: Operations,
+>(
     serialized: &[u8],
-) -> PolynomialRingElement {
+) -> PolynomialRingElement<Vector> {
     hax_debug_assert!(serialized.len() == (COEFFICIENTS_IN_RING_ELEMENT * COMPRESSION_FACTOR) / 8);
 
     match COMPRESSION_FACTOR as u32 {
@@ -267,37 +287,44 @@ pub(super) fn deserialize_then_decompress_ring_element_u<const COMPRESSION_FACTO
 }
 
 #[inline(always)]
-fn deserialize_then_decompress_4(serialized: &[u8]) -> PolynomialRingElement {
+fn deserialize_then_decompress_4<Vector: Operations>(
+    serialized: &[u8],
+) -> PolynomialRingElement<Vector> {
     hax_debug_assert!(serialized.len() == (COEFFICIENTS_IN_RING_ELEMENT * 4) / 8);
-    let mut re = PolynomialRingElement::ZERO();
+    let mut re = PolynomialRingElement::<Vector>::ZERO();
     cloop! {
         for (i, bytes) in serialized.chunks_exact(4).enumerate() {
-            let coefficient = simd::Vector::deserialize_4(&bytes);
-            re.coefficients[i] = simd::Vector::decompress::<4>(coefficient);
+            let coefficient = Vector::deserialize_4(&bytes);
+            re.coefficients[i] = Vector::decompress::<4>(coefficient);
         }
     }
     re
 }
 
 #[inline(always)]
-fn deserialize_then_decompress_5(serialized: &[u8]) -> PolynomialRingElement {
+fn deserialize_then_decompress_5<Vector: Operations>(
+    serialized: &[u8],
+) -> PolynomialRingElement<Vector> {
     hax_debug_assert!(serialized.len() == (COEFFICIENTS_IN_RING_ELEMENT * 5) / 8);
 
-    let mut re = PolynomialRingElement::ZERO();
+    let mut re = PolynomialRingElement::<Vector>::ZERO();
 
     cloop! {
         for (i, bytes) in serialized.chunks_exact(5).enumerate() {
-            re.coefficients[i] = simd::Vector::deserialize_5(&bytes);
-            re.coefficients[i] = simd::Vector::decompress::<5>(re.coefficients[i]);
+            re.coefficients[i] = Vector::deserialize_5(&bytes);
+            re.coefficients[i] = Vector::decompress::<5>(re.coefficients[i]);
         }
     }
     re
 }
 
 #[inline(always)]
-pub(super) fn deserialize_then_decompress_ring_element_v<const COMPRESSION_FACTOR: usize>(
+pub(super) fn deserialize_then_decompress_ring_element_v<
+    const COMPRESSION_FACTOR: usize,
+    Vector: Operations,
+>(
     serialized: &[u8],
-) -> PolynomialRingElement {
+) -> PolynomialRingElement<Vector> {
     hax_debug_assert!(serialized.len() == (COEFFICIENTS_IN_RING_ELEMENT * COMPRESSION_FACTOR) / 8);
 
     match COMPRESSION_FACTOR as u32 {
